allow only user to edit their own posts in node, init setup of userid in ng

+++ backend/routes/posts.js
@@ -33,7 +33,8 @@ router.post('', checkAuth, multer({storage: storage}).single('image'), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save()
     .then(result => {
@@ -61,9 +62,13 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: result.nModified + ' Document updated successful' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
+    .then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: result.nModified + ' Document updated successful' });
+      } else {
+        res.status(401).json({ message: 'Not authorized' });
+      }
   });
 });
 
@@ -105,10 +110,13 @@ router.get('/:id', (req, res) => {
 });
 
 router.delete('/:id', checkAuth, (req, res) => {
-  Post.deleteOne({ _id: req.params.id})
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId})
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: 'Post deleted' });
+      if (result.deletedCount > 0) {
+        res.status(200).json({ message: 'Post deleted' });
+      } else {
+        res.status(401).json({ message: 'Not authorized' });
+      }
     });
 });
